moved helpers to new file, tried to connect with related items, working on css
