+++ client/src/components/app.jsx
@@ -40,7 +40,7 @@ class App extends React.Component {
     return (
       <>
         <h3>Ateiler</h3>
-        <Overview />
+        <Overview outfitList={this.state.OutfitList} toggleStar={this.toggleStar} CurrentItemID={this.state.CurrentItemID} updateCurrentItem={this.updateCurrentItem}/>
         <RelatedItem outfitList={this.state.OutfitList} toggleStar={this.toggleStar} CurrentItemID={this.state.CurrentItemID} updateCurrentItem={this.updateCurrentItem} />
          <QandA product_name={this.state.CurrentItemName} product_id = {this.state.CurrentItemID}/>
         <RatingReview product_id={this.state.CurrentItemID} product_name={this.state.CurrentItemName} />
